+++ tests/spherical_harmonics_test.py
@@ -6,9 +6,13 @@ import pytest
 from jax.test_util import check_grads
 
 
-@pytest.fixture(scope="module", autouse=True, params=["legendre", "dense_tensor_product", "sparse_tensor_product"])
+@pytest.fixture(
+    scope="module",
+    autouse=True,
+    params=["legendre,sparse", "recursive,dense", "recursive,sparse,custom_vjp", "legendre,dense,custom_vjp"],
+)
 def algorithm(request):
-    e3nn.set_default_spherical_harmonics_algorithm(request.param)
+    e3nn.set_default_spherical_harmonics_algorithm(tuple(request.param.split(",")))
 
 
 @pytest.mark.parametrize("l", [0, 1, 2, 3, 4, 5, 6, 7])