+++ ChangeLog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `legendre` algorithm to compute spherical harmonics
+- add flag `algorithm` to specify the algorithm to use for computing spherical harmonics
 
 ## [0.6.1] - 2022-06-09
 ### Changed